move the databse models into their module
This is part of a restructure of how database connections/engines/ are created
